PP-10639: Move AgreementEvents to a new package
