+++ src/test/java/uk/gov/pay/connector/agreement/service/AgreementServiceTest.java
@@ -16,8 +16,8 @@ import uk.gov.pay.connector.agreement.model.AgreementEntity;
 import uk.gov.pay.connector.agreement.model.AgreementResponse;
 import uk.gov.pay.connector.charge.exception.PaymentInstrumentNotActiveException;
 import uk.gov.pay.connector.client.ledger.service.LedgerService;
-import uk.gov.pay.connector.events.model.charge.AgreementCancelledByService;
-import uk.gov.pay.connector.events.model.charge.AgreementCancelledByUser;
+import uk.gov.pay.connector.events.model.agreement.AgreementCancelledByService;
+import uk.gov.pay.connector.events.model.agreement.AgreementCancelledByUser;
 import uk.gov.pay.connector.gatewayaccount.dao.GatewayAccountDao;
 import uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntity;
 import uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntityFixture;
